feat: Complete External Merge of chunks
This implementation is somewhat interesting. I have not (and will not) go much further than this for now, but there are several things here other than the "merging" that might be very, very bad in real life situations.

First and foremost, a proper implementation would have quite a lot of I/O overhead, but this can be more or less waved away with modern systems and our puny amount of numbers.

Second, this would NOT run at all if we had to keep the entire script memory consumption under the set limit. I just did the minimum here to try to constrain the chunking and sorting itself, but not anything else.

Third, I am also pretty sure that this might not actually work for ludicrously large numbers (billion plus), but at that point it should be the concern of someone a lot higher up than this position is for.